Revert injection for east-west gateway
After testing recent changes for gateway injection, I found that injection worked for the ingress gateway, but not for the east-west gateway. Can't find anything in the Istio docs. Conferring w Ram next week, but for now removing the injection from the east-west gateway in , which worked in my tests.