Rm gloo_mesh_cluster_name
Remove gloo_mesh_cluster_name from istio installs in Gloo 2.1+, which is no longer required for metrics aggregation. Removed from all Gateway sample installs (Gloo Gateway is 2.1+ only), and from Mesh Istio 1.15 and 1.15 as these are only supported on 2.1+. For Istio 1.11-1.13 (which are supported for 2.0 and earlier as well as 2.1 and later), added a note that the field is only required in Mesh 2.0 and earlier.