Add istio 1.17, rm GG and GM 2.3
Add istio-operator files for 1.17. Technically in 2.3 and later we have instructions for manually installing w Helm, not IOP, but I'm adding them just in case customers want to continue using IOP and upgrade to 1.17.

Also removing the old-style values yamls for GG and GM in 2.3, since we are really pushing for profiles in the new chart in 2.3+